[refactor] group layout 생성
