+++ app/(main)/community/group/[groupId]/page.tsx
@@ -1,46 +1,15 @@
 'use client';
 
-import { PlusButton } from '@/components/atoms/Button';
-import Header from '@/components/atoms/Header';
 import Post from '@/components/molecules/Post';
-import { useRouter } from 'next/navigation';
 
 export default function GroupIdPage() {
-  const isMember = false;
-  const router = useRouter();
-
   return (
-    <div className='flex flex-col gap-[20px] w-full pt-[64px]'>
-      {/* 상단 헤더 */}
-      <div className='fixed top-0 left-0 right-0 z-10'>
-        <div className='mx-auto max-w-screen-md'>
-          <Header
-            text='꿈그룹'
-            showActionButton={true}
-            actionLabel={isMember ? '그룹 탈퇴' : '그룹 가입'}
-            actionTextColor={isMember ? 'primary-error' : 'primary-main'}
-          />
-        </div>
-      </div>
-
+    <div className='flex flex-col gap-[20px] w-full '>
       {/* <Header text='꿈그룹' showActionButton={false} /> */}
 
       {PostMockDatas.map((post) => (
         <Post key={post.id} {...post} />
       ))}
-
-      {/* 플러스 버튼 */}
-      {isMember && (
-        <div className='fixed left-0 right-0 bottom-[68px] max-w-screen-md mx-auto pointer-events-none'>
-          <PlusButton
-            className='absolute bottom-[20px] right-[20px] pointer-events-auto'
-            size='sm'
-            onClick={() => {
-              router.push('/community/create/post');
-            }}
-          />
-        </div>
-      )}
     </div>
   );
 }
