+++ app/(main)/community/group/[groupId]/post/page.tsx
@@ -1,3 +1,7 @@
 export default function PostPage() {
-  return <div>PostPage</div>;
+  return (
+    <div className='flex flex-col gap-[20px] w-full pt-[64px]'>
+      {/* 상단 헤더 */}
+    </div>
+  );
 }
